+++ yarn-project/cli/src/utils/aztec.ts
@@ -1,5 +1,6 @@
 import { type ContractArtifact, type FunctionArtifact, loadContractArtifact } from '@aztec/aztec.js/abi';
 import { type L1ContractArtifactsForDeployment } from '@aztec/aztec.js/ethereum';
+import { type DeployL1Contracts } from '@aztec/ethereum';
 import { type DebugLogger, type LogFn } from '@aztec/foundation/log';
 import { type NoirPackageConfig } from '@aztec/foundation/noir';
 import { GasTokenAddress } from '@aztec/protocol-contracts/gas-token';
@@ -33,17 +34,17 @@ export function getFunctionArtifact(artifact: ContractArtifact, fnName: string):
 /**
  * Function to execute the 'deployRollupContracts' command.
  * @param rpcUrl - The RPC URL of the ethereum node.
- * @param apiKey - The api key of the ethereum node endpoint.
+ * @param chainId - The chain ID of the L1 host.
  * @param privateKey - The private key to be used in contract deployment.
  * @param mnemonic - The mnemonic to be used in contract deployment.
  */
 export async function deployAztecContracts(
   rpcUrl: string,
-  apiKey: string,
-  privateKey: string,
+  chainId: number,
+  privateKey: string | undefined,
   mnemonic: string,
   debugLogger: DebugLogger,
-) {
+): Promise<DeployL1Contracts> {
   const {
     InboxAbi,
     InboxBytecode,
@@ -66,7 +67,7 @@ export async function deployAztecContracts(
   const account = !privateKey
     ? mnemonicToAccount(mnemonic!)
     : privateKeyToAccount(`${privateKey.startsWith('0x') ? '' : '0x'}${privateKey}` as `0x${string}`);
-  const chain = createEthereumChain(rpcUrl, apiKey);
+  const chain = createEthereumChain(rpcUrl, chainId);
   const l1Artifacts: L1ContractArtifactsForDeployment = {
     registry: {
       contractAbi: RegistryAbi,